requirements.txt
Django==5.2.3
djangorestframework==3.14.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3 